+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - ReleaseDate
 
+### Added
+
+- Keybinding for jj absorb (`A`)
+
 ## [0.8.0] - 2026-04-19
 
 ### Added